Never check a version before installing
It's no longer feasible to check if a version
exists before installing because we can no longer determine the correct
location based on the version number, and Docker no longer tags all of
its releases. So the only way to check if a version exists is to check
multiple locations for the existance of a file, so may as well just do
the install at that point.